perf(render): sublayer refactor & optimizations for series rendering
