+++ src/FastCharts.Rendering.Skia/Rendering/Layers/ISeriesSubLayer.cs
@@ -1,0 +1,7 @@
+namespace FastCharts.Rendering.Skia.Rendering.Layers
+{
+    internal interface ISeriesSubLayer
+    {
+        void Render(RenderContext ctx);
+    }
+}
