+++ src/FastCharts.Rendering.Skia/Rendering/Layers/LineLayer.cs
@@ -1,0 +1,32 @@
+using SkiaSharp;
+using FastCharts.Core.Series;
+
+namespace FastCharts.Rendering.Skia.Rendering.Layers
+{
+    internal sealed class LineLayer : ISeriesSubLayer
+    {
+        public void Render(RenderContext ctx)
+        {
+            var model = ctx.Model; var pr = ctx.PlotRect; var palette = model.Theme.SeriesPalette; int paletteCount = palette?.Count ?? 0;
+            int lineIndex = 0;
+            foreach (var s in model.Series)
+            {
+                if (s is not LineSeries ls) continue;
+                if (ls is AreaSeries || ls is ScatterSeries || ls is StepLineSeries) continue;
+                if (ls.IsEmpty || !ls.IsVisible) continue;
+                var c = (paletteCount > 0 && lineIndex < paletteCount) ? palette[lineIndex] : model.Theme.PrimarySeriesColor;
+                using var sp = new SKPaint { IsAntialias = true, Style = SKPaintStyle.Stroke, StrokeWidth = (float)ls.StrokeThickness, Color = new SKColor(c.R, c.G, c.B, c.A) };
+                using var path2 = new SKPath(); bool startedLine = false;
+                foreach (var p in ls.Data)
+                {
+                    float px = PixelMapper.X(p.X, model.XAxis, pr);
+                    float py = PixelMapper.Y(p.Y, model.YAxis, pr);
+                    if (!startedLine) { path2.MoveTo(px, py); startedLine = true; }
+                    else { path2.LineTo(px, py); }
+                }
+                ctx.Canvas.Save(); ctx.Canvas.ClipRect(pr); ctx.Canvas.DrawPath(path2, sp); ctx.Canvas.Restore();
+                lineIndex++;
+            }
+        }
+    }
+}
